% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning_numbers.R
\name{binning_numbers}
\alias{binning_numbers}
\title{Binning numbers function}
\usage{
binning_numbers(sequence_to_cut, desired_bin_size, right = FALSE)
}
\arguments{
\item{sequence_to_cut}{A sequence of numbers that you want to know how many numbers are in however many bins}

\item{desired_bin_size}{The bin size you want}

\item{right}{default is FALSE}
}
\value{
a frequency table
}
\description{
Binning numbers function
}
\examples{
sequence_to_cut <- 0:100
binning_numbers(sequence_to_cut, 5)
}
