% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_plot.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Histogram plot}
\usage{
histogram_plot(sequence_to_cut, desired_bin_size)
}
\arguments{
\item{sequence_to_cut}{the sequence to be binned}

\item{desired_bin_size}{how big the bins should be}
}
\value{
ggplot object that can be plotted
}
\description{
Histogram plot
}
